requirements.txt
fastapi
uvicorn
pytesseract
pillow
pdfplumber